refactor: split remote tests.rs into handshake, relay, and lifecycle modules
Break the 1028-line monolithic test file into three focused submodules:
- handshake: protocol serialization, key validation, connection state
- relay: message sending, replay buffer behavior
- lifecycle: registration, heartbeat, reap/stale, metrics, reconnect, subscriptions

All 46 tests pass with clean clippy under strict warnings.
